1 服务器拍卖行列表每日存储和刷新,每天存储当前的购买数量 2 玩家购买当前道具,会从拍卖行找价格低的玩家进行购买,

+++ Server/Model/Generate/Message/OuterMessage.cs
@@ -5125,7 +5125,7 @@ namespace ET
 		public float PricePro { get; set; }
 
 		[ProtoMember(6)]
-		public float BuyNum { get; set; }
+		public int BuyNum { get; set; }
 
 	}
